Move DialogTitleLayout to the main subpackage
