+++ core/src/main/java/com/afollestad/materialdialogs/internal/main/DialogLayout.kt
@@ -23,7 +23,6 @@ import androidx.annotation.ColorInt
 import com.afollestad.materialdialogs.MaterialDialog
 import com.afollestad.materialdialogs.R
 import com.afollestad.materialdialogs.internal.button.DialogActionButtonLayout
-import com.afollestad.materialdialogs.internal.title.DialogTitleLayout
 import com.afollestad.materialdialogs.utils.dimenPx
 
 val DEBUG_COLOR_PINK = Color.parseColor("#EAA3CF")
