+++ core/src/main/res/layout/md_dialog_base.xml
@@ -7,7 +7,7 @@
     android:layout_height="wrap_content"
     >
 
-  <com.afollestad.materialdialogs.internal.title.DialogTitleLayout
+  <com.afollestad.materialdialogs.internal.main.DialogTitleLayout
       android:id="@+id/md_title_layout"
       android:layout_width="match_parent"
       android:layout_height="wrap_content"
@@ -31,7 +31,7 @@
         style="@style/MD_Dialog_Title_Text"
         />
 
-  </com.afollestad.materialdialogs.internal.title.DialogTitleLayout>
+  </com.afollestad.materialdialogs.internal.main.DialogTitleLayout>
 
   <com.afollestad.materialdialogs.internal.button.DialogActionButtonLayout
       android:id="@+id/md_button_layout"
